style: format code with forge fmt

+++ test/Solinference.t.sol
@@ -17,8 +17,8 @@ contract SolinferenceTest is Test {
         solinference = new Solinference(address(zTableInstance));
     }
 
-    function testMean() public view{
-        int[] memory data = new int[](41);
+    function testMean() public view {
+        int256[] memory data = new int256[](41);
         data[0] = 1e18; // 1
         data[1] = 2e18; // 2
         data[2] = 3e18; // 3
@@ -59,13 +59,13 @@ contract SolinferenceTest is Test {
         data[37] = 37e18; // 37
         data[38] = 38e18; // 38
         data[39] = 39e18; // 39
-        data[40] = 40e18; // 40 
+        data[40] = 40e18; // 40
         int256 result = solinference.mean(data);
-        assertApproxEqAbs(result, 20243902439024390244, 1e12);  // Expected mean is 20.24390243902439
+        assertApproxEqAbs(result, 20243902439024390244, 1e12); // Expected mean is 20.24390243902439
     }
 
-    function testMedian() public view{
-        int[] memory data = new int[](41);
+    function testMedian() public view {
+        int256[] memory data = new int256[](41);
         data[0] = 1e18; // 1
         data[1] = 2e18; // 2
         data[2] = 3e18; // 3
@@ -106,13 +106,13 @@ contract SolinferenceTest is Test {
         data[37] = 37e18; // 37
         data[38] = 38e18; // 38
         data[39] = 39e18; // 39
-        data[40] = 40e18; // 40 
+        data[40] = 40e18; // 40
         int256 result = solinference.median(data);
         assertApproxEqAbs(result, 20e18, 1e12); // Expected median is 20
     }
 
     function testRange() public view {
-        int[] memory data = new int[](41);
+        int256[] memory data = new int256[](41);
         data[0] = 1e18; // 1
         data[1] = 2e18; // 2
         data[2] = 3e18; // 3
@@ -153,13 +153,13 @@ contract SolinferenceTest is Test {
         data[37] = 37e18; // 37
         data[38] = 38e18; // 38
         data[39] = 39e18; // 39
-        data[40] = 40e18; // 40 
-        uint256 result = solinference.range(data); 
-        assertApproxEqAbs(result, 39e18, 1e12);  // Expected range is 40 - 1 = 39
+        data[40] = 40e18; // 40
+        uint256 result = solinference.range(data);
+        assertApproxEqAbs(result, 39e18, 1e12); // Expected range is 40 - 1 = 39
     }
 
     function testMode() public view {
-        int[] memory data = new int[](41);
+        int256[] memory data = new int256[](41);
         data[0] = 1e18; // 1
         data[1] = 2e18; // 2
         data[2] = 3e18; // 3
@@ -200,14 +200,14 @@ contract SolinferenceTest is Test {
         data[37] = 37e18; // 37
         data[38] = 38e18; // 38
         data[39] = 39e18; // 39
-        data[40] = 40e18; // 40 
-        int[] memory result = solinference.mode(data);
+        data[40] = 40e18; // 40
+        int256[] memory result = solinference.mode(data);
         assertEq(result.length, 1);
         assertEq(result[0], 10e18); // Expected mode is 10
     }
 
     function testVariance() public view {
-        int[] memory data = new int[](41);
+        int256[] memory data = new int256[](41);
         data[0] = 1e18; // 1
         data[1] = 2e18; // 2
         data[2] = 3e18; // 3
@@ -248,15 +248,13 @@ contract SolinferenceTest is Test {
         data[37] = 37e18; // 37
         data[38] = 38e18; // 38
         data[39] = 39e18; // 39
-        data[40] = 40e18; // 40 
+        data[40] = 40e18; // 40
         int256 result = solinference.variance(data);
         assertApproxEqAbs(result, 132623438429506240000, 1e6); // Adjusted expected value and tolerance
     }
-   
-
 
     function testStdDev() public view {
-        int[] memory data = new int[](41);
+        int256[] memory data = new int256[](41);
         data[0] = 1e18; // 1
         data[1] = 2e18; // 2
         data[2] = 3e18; // 3
@@ -297,14 +295,13 @@ contract SolinferenceTest is Test {
         data[37] = 37e18; // 37
         data[38] = 38e18; // 38
         data[39] = 39e18; // 39
-        data[40] = 40e18; // 40 
+        data[40] = 40e18; // 40
         int256 result = solinference.stdDev(data);
         assertApproxEqAbs(result, 11516225007766488000, 1e10); // Expected std deviation is 11.516225007766488
-    } 
-
+    }
 
     function testZScore() public view {
-        int[] memory data = new int[](41);
+        int256[] memory data = new int256[](41);
         data[0] = 1e18; // 1
         data[1] = 2e18; // 2
         data[2] = 3e18; // 3
@@ -345,14 +342,13 @@ contract SolinferenceTest is Test {
         data[37] = 37e18; // 37
         data[38] = 38e18; // 38
         data[39] = 39e18; // 39
-        data[40] = 40e18; // 40 
-        int result = solinference.zScore(data, 25e18);
-        assertEq(result, 264); 
+        data[40] = 40e18; // 40
+        int256 result = solinference.zScore(data, 25e18);
+        assertEq(result, 264);
     }
-    
 
     function testGetProbability() public view {
-        int[] memory data = new int[](41);
+        int256[] memory data = new int256[](41);
         data[0] = 1e18; // 1
         data[1] = 2e18; // 2
         data[2] = 3e18; // 3
@@ -393,19 +389,19 @@ contract SolinferenceTest is Test {
         data[37] = 37e18; // 37
         data[38] = 38e18; // 38
         data[39] = 39e18; // 39
-        data[40] = 40e18; // 40 
-        
+        data[40] = 40e18; // 40
+
         int256 meanValue = solinference.mean(data);
         int256 stdDevValue = solinference.stdDev(data);
         int256 zScoreValue = solinference.zScore(data, 25e18);
-        
+
         console.log("Mean:", uint256(meanValue));
         console.log("StdDev:", uint256(stdDevValue));
         console.log("Z-Score:", uint256(zScoreValue));
-        
+
         int256 result = solinference.getProbability(data, 25e18);
         console.log("Probability result:", uint256(result));
-        
+
         // The z-table values are in basis points (e.g., 9959 = 99.59%)
         // So we should expect a value from the z-table mapping
         assertEq(result, zTableInstance.zTable(zScoreValue));
